modify: ProductPost init, 판매상품 - 상세페이지 연결, 판매상품 거래완료-거래중 구분
1. PruductPost init할 떄 [Post]타입의 데이터가 전달되도록 수정
2. 판매상품 - 상세페이지 화면 연결 및 데이터 전달
3. 판매상품페이지 - 전체, 거래중, 거래완료 상품 구분해서 정렬
