+++ DaangnMarket/DaangnMarket/MyDaangn/MyPageViewController.swift
@@ -10,6 +10,7 @@ import UIKit
 
 class MyPageViewController: UIViewController {
   override func viewDidLoad() {
+    view.backgroundColor = .white
     super.viewDidLoad()
   }
 }
