+++ DaangnMarket/DaangnMarket/MyDaangn/ProfilePage/View/UserProfilePercentageView.swift
@@ -16,8 +16,8 @@ class UserProfilePercentageView: UIView {
   // MARK: Views
   
   private let heartMark = UIButton().then {
-   $0.setImage(UIImage(systemName: "heart.fill"), for: .normal)
-   $0.tintColor = .systemGreen  // 색상 추가!!!!!!!!
+    $0.setImage(UIImage(systemName: "heart.fill"), for: .normal)
+    $0.tintColor = .systemGreen  // 색상 추가!!!!!!!!
     $0.contentMode = .scaleAspectFit
   }
   private let hopingRebuyingLabel = UILabel().then {
@@ -30,10 +30,10 @@ class UserProfilePercentageView: UIView {
     $0.attributedText = title
   }
   private let hopingRebuyingExplainLabel = UILabel().then {
-   $0.numberOfLines = 0
-   $0.text = "표시될 만큼 충분히 거래하지 않았어요."
-   $0.textColor = UIColor(named: ColorReference.subText.rawValue)
-   $0.font = UIFont.systemFont(ofSize: 13)
+    $0.numberOfLines = 0
+    $0.text = "표시될 만큼 충분히 거래하지 않았어요."
+    $0.textColor = UIColor(named: ColorReference.subText.rawValue)
+    $0.font = UIFont.systemFont(ofSize: 13)
   }
   private let storyBubbleMark = UIButton().then {
     $0.setImage(UIImage(systemName: "bubble.right.fill"), for: .normal)
@@ -41,9 +41,9 @@ class UserProfilePercentageView: UIView {
     $0.contentMode = .scaleAspectFit
   }
   private let chatResponseLabel = UILabel().then {
-   $0.textColor = .black
-   $0.font = UIFont.boldSystemFont(ofSize: 14)
-   $0.text = "응답률 -%"
+    $0.textColor = .black
+    $0.font = UIFont.boldSystemFont(ofSize: 14)
+    $0.text = "응답률 -%"
     let fontSize = UIFont.systemFont(ofSize: 13)
     let title = NSMutableAttributedString(string: $0.text!)
     title.addAttribute(NSAttributedString.Key(rawValue: kCTFontAttributeName as String), value: fontSize, range: ($0.text! as NSString).range(of: "응답률"))
